Calibration to double

+++ camera_stabilization/cpp/lib/PerspectiveProjection.cpp
@@ -6,13 +6,13 @@
 
 namespace providentia {
 	namespace camera {
-		Eigen::Vector4f normalize(const Eigen::Vector4f &vector) {
+		Eigen::Vector4d normalize(const Eigen::Vector4d &vector) {
 			return vector / vector(3);
 		}
 
-		PerspectiveProjection::PerspectiveProjection(float sensorWidth, float aspectRatio, float focalLength,
-													 float nearPlaneDistance,
-													 float farPlaneDistance) {
+		PerspectiveProjection::PerspectiveProjection(double sensorWidth, double aspectRatio, double focalLength,
+													 double nearPlaneDistance,
+													 double farPlaneDistance) {
 			setFieldOfView(sensorWidth, aspectRatio, focalLength);
 			setFrustumPlaneDistances(nearPlaneDistance, farPlaneDistance);
 			setImagePlaneBounds(nearPlaneDistance, aspectRatio);
@@ -39,18 +39,18 @@ namespace providentia {
 			projection = normalization * frustum;
 		}
 
-		void PerspectiveProjection::setFrustumPlaneDistances(float nearPlaneDistance,
-															 float farPlaneDistance) {
+		void PerspectiveProjection::setFrustumPlaneDistances(double nearPlaneDistance,
+															 double farPlaneDistance) {
 			frustumPlaneDistances = {nearPlaneDistance, farPlaneDistance};
 		}
 
-		void PerspectiveProjection::setImagePlaneBounds(float nearPlaneDistance, float aspectRatio) {
+		void PerspectiveProjection::setImagePlaneBounds(double nearPlaneDistance, double aspectRatio) {
 			topRight = {nearPlaneDistance * tan(fieldOfView.x() * 0.5), 1};
 			topRight.y() = topRight.x() / aspectRatio;
 			lowerLeft = -topRight;
 		}
 
-		void PerspectiveProjection::setFieldOfView(float sensorWidth, float aspectRatio, float focalLength) {
+		void PerspectiveProjection::setFieldOfView(double sensorWidth, double aspectRatio, double focalLength) {
 			fieldOfView = {2. * atan(0.5 * (sensorWidth / focalLength)), 1};
 			fieldOfView.y() = fieldOfView.x() / aspectRatio;
 		}
@@ -62,23 +62,23 @@ namespace providentia {
 			return os;
 		}
 
-		Eigen::Vector2f PerspectiveProjection::operator*(const Eigen::Vector4f &vectorInCameraSpace) {
+		Eigen::Vector2d PerspectiveProjection::operator*(const Eigen::Vector4d &vectorInCameraSpace) {
 			return toClipSpace(vectorInCameraSpace).head<2>();
 		}
 
-		Eigen::Vector4f PerspectiveProjection::toFrustum(const Eigen::Vector4f &vectorInCameraSpace) {
+		Eigen::Vector4d PerspectiveProjection::toFrustum(const Eigen::Vector4d &vectorInCameraSpace) {
 			return normalize(frustum * vectorInCameraSpace);
 		}
 
-		Eigen::Vector3f PerspectiveProjection::toClipSpace(const Eigen::Vector4f &vectorInCameraSpace) {
+		Eigen::Vector3d PerspectiveProjection::toClipSpace(const Eigen::Vector4d &vectorInCameraSpace) {
 			return normalize(projection * vectorInCameraSpace).head<3>();
 		}
 
-		const Eigen::Vector2f &PerspectiveProjection::getTopRight() const {
+		const Eigen::Vector2d &PerspectiveProjection::getTopRight() const {
 			return topRight;
 		}
 
-		const Eigen::Vector2f &PerspectiveProjection::getLowerLeft() const {
+		const Eigen::Vector2d &PerspectiveProjection::getLowerLeft() const {
 			return lowerLeft;
 		}
